fastapi-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
aiohttp
python-nmap
libnmap
supabase
openai
python-dotenv
